% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_traversal_util.R
\name{calculateAMDMS}
\alias{calculateAMDMS}
\title{Forest Averaged Minimal Depth of a Maximal Subtree (AMDMS)}
\usage{
calculateAMDMS(ranger_obj)
}
\arguments{
\item{ranger_obj}{A ranger object from the ranger package, which was created
with param write.forest set to TRUE. In other words, it must have a 
'forest' property.}
}
\value{
A data.frame with two columns: averaged first and second order 
minimal depth of a maximal subtree.
}
\description{
Given a result from the Ranger package (write.forest must
be set to TRUE), this function will traverse the trees and calculate the
first and second order average minimal depth of a maximal subtree.
}

