% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_traversal_util.R
\name{startRecursiveDepthBinning}
\alias{startRecursiveDepthBinning}
\title{Start Recursive Depth Binning}
\usage{
startRecursiveDepthBinning(tree_split_varIDs)
}
\arguments{
\item{tree_split_varIDs}{Given one element of a 'split.varIDs' list, this
function will pass it to the recursiveDepthBinning function to bin the tree
by depth, starting at the root.}
}
\value{
A list with an element per depth encountered. Each 
element is a vector of variable IDs
}
\description{
The starter function for the recursion in recursiveDepthBinning.
}

